refactor: bag 2가지 이미지 분류(fill, outline)
