+++ src/components/Navbar/Navbar.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
-import menu from '../../assets/images/bag.png';
+import menu from '../../assets/images/bag-outline.png';
 
 const NavLayout = styled.header`
   width: 100%;
@@ -40,6 +40,7 @@ const NavMenu = styled.li`
 `;
 
 const NavMenuIcon = styled.img`
+  width: 1.5rem;
   margin-right: 0.3rem;
 `;
