+++ src/pages/Room.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import Bestlocation from '../components/Card/Bestlocation';
-import Recreation from '../components/Card/Recreation';
+import Recreation from '../components/Card/recreation';
 import Bag from '../components/Card/Bag';
 
 const Room = () => {
